requirements_runpod.txt
# CP-VTON+ Requirements for RunPod Serverless
# استخدام PyTorch أحدث من الأصلي (0.4.1) للتوافق مع CUDA الحديث

# RunPod SDK
runpod>=1.6.0

# PyTorch & Vision
torch>=1.10.0,<2.0.0
torchvision>=0.11.0,<0.15.0

# Image processing
Pillow>=8.0.0
opencv-contrib-python>=4.5.0

# Utilities
numpy>=1.19.0

# Optional: للـ tensorboard (لو حابب تستخدمها)
tensorboardX>=2.2

